backend/requirements.txt
fastapi==0.124.4
uvicorn[standard]==0.38.0
python-multipart==0.0.20
python-dotenv==1.2.1
pydantic-settings==2.12.0
aiofiles==25.1.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.30.0
psycopg2-binary==2.9.9
alembic==1.13.1
pgvector==0.4.2
# Auth
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
email-validator==2.1.1

# Email
fastapi-mail==1.4.1

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.27.0

pytest-cov==4.1.0        # Coverage reporting
pytest-mock==3.12.0      # Easier mocking